LPD-90495 updates RestController classes to use updated services
