+++ workspaces/liferay-one-workspace/client-extensions/liferay-one-etc-spring-boot/src/main/java/com/liferay/one/TicketAttachmentsRestController.java
@@ -15,7 +15,7 @@ import com.liferay.one.jira.constants.IssueConstants;
 import com.liferay.one.jira.exception.OrganizationNotFoundException;
 import com.liferay.one.jira.model.Organization;
 import com.liferay.one.jira.model.SupportIssue;
-import com.liferay.one.jira.service.JiraService;
+import com.liferay.one.jira.service.JiraIssueService;
 import com.liferay.one.model.TicketAttachment;
 import com.liferay.one.service.GoogleCloudStorageService;
 import com.liferay.one.service.NotificationQueueEntryService;
@@ -219,7 +219,7 @@ public class TicketAttachmentsRestController extends OneBaseRestController {
 			jsonObject.optString("commentBody"), jwt, ticketAttachment);
 
 		try {
-			_jiraService.addComment(
+			_jiraIssueService.addComment(
 				jiraIssueCommentBody, ticketAttachment.getJiraIssueKey());
 
 			return new ResponseEntity<>(HttpStatus.OK);
@@ -251,7 +251,7 @@ public class TicketAttachmentsRestController extends OneBaseRestController {
 
 		String ticketId = jsonObject.getString("ticketId");
 
-		SupportIssue supportIssue = _jiraService.getSupportIssue(ticketId);
+		SupportIssue supportIssue = _jiraIssueService.getSupportIssue(ticketId);
 
 		if (supportIssue == null) {
 			return new ResponseEntity<>(
@@ -326,7 +326,7 @@ public class TicketAttachmentsRestController extends OneBaseRestController {
 		sb.append(StringUtil.merge(IssueConstants.STATUSES_CLOSED, "', '"));
 		sb.append("') after -8d before -7d)");
 
-		List<SupportIssue> supportIssues = _jiraService.search(
+		List<SupportIssue> supportIssues = _jiraIssueService.search(
 			sb.toString(), new String[] {"key"});
 
 		for (SupportIssue supportIssue : supportIssues) {
@@ -378,7 +378,7 @@ public class TicketAttachmentsRestController extends OneBaseRestController {
 
 		for (TicketAttachment ticketAttachment : ticketAttachments) {
 			try {
-				_jiraService.addComment(
+				_jiraIssueService.addComment(
 					ticketAttachment.getDraftCommentBody(),
 					ticketAttachment.getJiraIssueKey());
 
@@ -636,7 +636,7 @@ public class TicketAttachmentsRestController extends OneBaseRestController {
 	private GoogleCloudStorageService _googleCloudStorageService;
 
 	@Autowired
-	private JiraService _jiraService;
+	private JiraIssueService _jiraIssueService;
 
 	@Value("${liferay.one.jira.support.fls.project}")
 	private String _jiraSupportFLSProject;
